ipam: Move metrics and stats to `operator/pkg`
`pkg/ipam/metrics` and `pkg/ipam/stats` are only imported by operator
code.

This commit continues work started in afbcb75fae08 and moves those
packages under `operator/pkg/ipam/` so the agent/operator
boundary becomes more clear.
